Don't parse stderr for errors, rely on exit code check
Following the execution of external commands, such as pg_dump,
pg_dumpall, psql, dropdb, createdb, stderr output is no longer
parsed for ERROR or FATAL messages. Instead we rely on checking
the exit status of those commands.

The change is made because stderr could legitimately contain the
characters 'ERROR' or 'FATAL' as part of successful
execution of the external commands, leading to false failures.

When the checks of stderr was first added, they were the only test for
error during execution (though they didn't actually trap all error
conditions). Now that the process' exit code is checked, it is not
necessary to parse stderr.

* `psql` is now run with the `--set=ON_ERROR_STOP=on` switch.
* `pg_restore` is now run with the `--exit-on-error` switch.