Make file tests depend on DB_TESING
Renamed the old `03-shellexceptions tests.t` to `03-file_errors.t`.

We no longer involve the shell in running external commands (since
0.120.0) so the previous test name was misleading.

Now that we no longer use the shell, the response to file errors
comes from the underlying postgreSQL utility (such as pg_dump). We
need these to be available for the tests to run, so the test is
now dependent on DB_TESTING environment variable being set.